fix(tools): address code-review findings on today's two commits
Critical:
- AuditFamiliesTool: ElementId equality must use .Equals() — on net48 the
  == operator is reference-equality and silently excludes every family
  when categoryFilter is set. R25 was not affected (operator overload
  present since .NET ElementId.Value overload).
- AuditFamiliesTool: OfClass(typeof(RoofBase)) throws ArgumentException
  because RoofBase is abstract and not a supported collector class.
  Switched the system-family instance count to OfCategory(OST_*) which
  captures FootPrintRoof, ExtrusionRoof (and Wall/Floor/Ceiling analogs)
  in one pass and doesn't need a concrete instance class.

Should-fix:
- ManageProjectParametersTool: empty CategorySet post-build is now
  rejected with InvalidInput (was masked as PermissionDenied by the
  ReInsert-false guard).
- CategoryResolver: dropped the IndexOf partial-match fallback — "Stair"
  was silently matching "Stair Support" and friends. Resolution is now
  strictly exact match (OST_ → friendly EN → localized display name).
- FilterByParameterValueTool: pre-validate doc.ActiveView before using
  scope="active_view" — prevents NRE when requests arrive without UI
  context.

Nice-to-have:
- All server-side element ID parameters normalized to long / long[]
  (get_element_parameters, operate_element, section_box_from_selection,
  create_floor, export_schedule, manage_links, compound-structure,
  create_room, duplicate_system_type, duplicate/place viewport, view
  and schedule ops, filter helpers). int[] truncated to 32-bit on
  R2024+ at the server boundary.
- AuditFamiliesTool: system-family synthetic row id changed from 0 to
  -1 to match the null-sentinel convention used elsewhere.

+++ src/RevitCortex.Server/Tools/CreationTools.cs
@@ -45,7 +45,7 @@ public static class CreationTools
     [McpServerTool(Name = "create_floor"), Description("Create a floor element from boundary points, type name, and level.")]
     public static async Task<string> CreateFloor(
         RevitConnectionManager revit,
-        [Description("Level element ID")] int levelId,
+        [Description("Level element ID")] long levelId,
         [Description("JSON array of boundary points [{x, y, z}]")] string boundaryPoints,
         [Description("Floor type name")] string? typeName = null,
         CancellationToken ct = default)
@@ -64,7 +64,7 @@ public static class CreationTools
     public static async Task<string> ChangeElementType(
         RevitConnectionManager revit,
         [Description("Element IDs to change")] int[] elementIds,
-        [Description("Target type element ID")] int? targetTypeId = null,
+        [Description("Target type element ID")] long? targetTypeId = null,
         [Description("Target type name")] string? targetTypeName = null,
         CancellationToken ct = default)
     {
@@ -116,7 +116,7 @@ public static class CreationTools
         RevitConnectionManager revit,
         [Description("Category name (e.g. Walls, Doors)")] string category,
         [Description("Color as hex string (e.g. #FF0000)")] string color,
-        [Description("View element ID. Uses active view if not specified")] int? viewId = null,
+        [Description("View element ID. Uses active view if not specified")] long? viewId = null,
         CancellationToken ct = default)
     {
         var p = new JObject
@@ -346,7 +346,7 @@ public static class CreationTools
     [McpServerTool(Name = "export_schedule"), Description("Export schedule to CSV/TSV or JSON")]
     public static async Task<string> ExportSchedule(
         RevitConnectionManager revit,
-        [Description("Schedule element ID")] int scheduleId,
+        [Description("Schedule element ID")] long scheduleId,
         [Description("Export format (csv, tsv, json). Default: csv")] string? format = "csv",
         CancellationToken ct = default)
     {
